Move 3rdparty/{email,sms}-verification to correct location
