+++ js/src/views/Account/Verification/GatherData/gatherData.js
@@ -24,8 +24,8 @@ import { Checkbox, Form, Input } from '@parity/ui';
 import { DoneIcon, ErrorIcon, InfoIcon } from '@parity/ui/Icons';
 import { nullableProptype } from '@parity/shared/util/proptypes';
 
-import smsTermsOfService from '~/3rdparty/sms-verification/terms-of-service';
-import emailTermsOfService from '~/3rdparty/email-verification/terms-of-service';
+import smsTermsOfService from '~./sms-verification/terms-of-service';
+import emailTermsOfService from './email-verification/terms-of-service';
 
 import { howSMSVerificationWorks, howEmailVerificationWorks } from '../how-it-works';
 import styles from './gatherData.css';
