+++ js/src/views/Account/Verification/email-verification/index.js
