+++ js/src/views/Account/Verification/email-verification/terms-of-service.js
