+++ js/src/views/Account/Verification/sms-verification/terms-of-service.js
